(#322) Added skip logic to the numeric steps.
The runtime part it's also implemented. If the reply value isn't empty
but it doesn't find a valid range it will default to the next question.
The numeric step fields min_value, max_value and ranges_delimiters are only used by
the UI, the runtime only needs ranges.
The UI won't update the ranges if the description of them by the user
(min_value, max_value and ranges_delimiters) is invalid. It also
remembers the skip logic option chosen for a range if it doesn't change
on a range edit.