Fixes #443
Most of the functionality for #443 was already done in previous commits. This last commit focuses on enhancing the SkipLogic component to make it react properly to current invalid selections. The component now shows a striked through text in red when the current value points at a step that is no longer after the current step (or even that was removed from the questionnaire altogether).