Fixes #752. Both primary/fallback mode show IVR when running a mode comparison survey
Until this point, only a primary and a fallback mode was being displayed in the Survey Show
screen.

This is not correct, because you can actually set a comparison between modes when
creating the survey.

To fix that, a primary and (if needed) a fallback mode is displayed for each mode selected
for the comparison.