For #885. Reverting changes related with font size
Font size will be the same for all prompts, despite their lengths,
as it was implemented from the beginning.

Usually steps have short prompts, with some exceptions like: Intro &
consent steps. In that case, it is better for the user to scroll down
instead of reducing the font size.

Commit includes some little changes added to button font-sizes classes