Fixes #1024. Step disappears while dragging
The bug is induced by materializecss. It applies a property left: -9999px
to checkboxes and radio buttons, contained in numeric and flag steps
respectively. That left property is what produce the bug. That is,
drag and drop failing for that kind of steps.

In order to avoid any kind of hack, we took the decision of
remove drag and drop for step editors (a.k.a, expanded steps).