Fix validations in questionnaireEditor for MultiLanguage without sections
When there is no sections in the multiple language 
case, 2 separated groupSteps are rendered: the language selection step on the one hand and the main group step on the other.

Errors for the steps of the main group are shifted by one position with respect to the local index on the groupStep (because the language selection step group is rendered separately).