Closes #1426. Make all audio URL's to end in ".mp3"
Phoenix doesn't provide a straightforward way to append an ".mp3" extension to the routes in router.ex.
For that reason, the ".mp3" extension is added manually in verboice_channel.ex, and an :mp3_api pipeline is added in router.ex. The pipeline accepts "mp3" format, and uses the TrailingFormatPlug library to properly handle request that finish with ".mp3" extension.