Fixes #1425. Quantity totals are not right when quotas are configured
The count for the respondents that didn't have an assigned quota bucket wasn't being displayed. The total quantity, however, was being computed properly.

The chosen solution is to display a new group: 'Unassigned', which groups the respondents without quota buckets. Empty string '' is used as the referenceId for that purpose, which is already being computed in the backend.

Along with the fix, two additional changes are included. The Zero value is replaced with a dash: "-", and 'registered' respondents' count is shown again to keep the values of the table consistent.