Closes #1443. Autosave time is displayed incorrectly
The '+0000' suffix was being appended to saveStatus.updatedAt in order to provide the UTC offset  (commit 5ad8536). 

However, since commit b825a12 that suffix is not necessary anymore, because updated_at field changed from NaiveDateTime to DateTime, which already contains the timezone info.

The solution to the current issue is to remove the '+0000' suffix because it was providing duplicated info, leading to parsing errors of the datetime.