#1449. Interrupt survey processing when a channel is down.
If a single channel is down, the broker has to stop polling that survey. Even though the other channels are :up.
Additionally, receiving :unknown means that the ChannelStatusServer doesn't know the status of the channel, and in that case the broker assumes that the channel is :up.
Finally, ChannelStatusServer is scoped to Ask.Runtime instead of Ask to keep naming consistency because it is used at runtime.