Accumulate calls durations to calculate total call time
Surveda receives call durations on Verboice status callbacks.

Surveda was keeping the latest call duration received as the total
call durations for the respondent, which was wrong. Now we keep
record of each of the calls involved in a respondent's survey, so
we can then sum them up.

Surveda can receive multiple status callbacks for a single call, so
it would be difficult to sum that in a single variable.

See #1452