Unify respondent locks identifiers
Depending on the caller of the `Respondent.with_lock` function, the
respondent_id paremeter was either an Integer or String, so the
lock wasn't useful when Broker and Channels callbacks were competing.

We now unify them as strings so they compite for the same lock.

We still enqueue Verboice calls 5 seconds in the future to avoid the
callbacks to timeout (see #1623).

See #1737

Co-authored-by: Ana Pérez Ghiglia <aghiglia@manas.tech>