Avoid future behaviour misunderstanding
In previous iterations of the panel survey development, the new wave
took its questionnaires from the latest wave of its panel survey.

Currently, the end-user select a questionnaire after the new wave is
created.

The lines this commit removes can only lead to misunderstanding.
Because of them, the new wave could wrongly appear (to future Devs or mantainers) to be associated with a questionnaire.

But really it isn't because the UI doesn't show the wave <-> questionnaire
association to the end-user. Indeed, it forces them to select a
questionnaire.