Fix: simulation ended toast can be displayed multiple times (#1980)
Mark a simulation has ended when it expired or ended, so we can
never display the expired/ended toast multiple times. This didn't
happen for the SMS and IVR simulators, but the mobileweb one runs in
its own iframe, and can send a message multiple times.

Refactors to only listen to the `message` event for the mobileweb
simulator, and removes the event manually added to `window` when the
component is unmounted, otherwise we'd have a memory leak with
unmounted components continuing to receive the mobileweb simulator
messages, leading to errors in the console, as well as memory that
the Garbage Collector would never reclaim.