Sandbox: add chat history to IVR simulator responses (#2006)
* API: add messages_history to IVR simulator responses
* Extract MessagesList component from ChatWindow
* Add MessagesList to VoiceWindow component

That needed to change how the HTML component spans in the available
vertical space. It used to be a fixed height, but now we use flexbox
to let it span the available, dynamic space (there is lots of space
in the SMS simulator, but not that much in the IVR simulator).