Fix: performance issue in RespondentGroupAction.remove_duplicates_with_respect_to/2
Adds an index to respondents.canonical_phone_numbers because we use
it to determine possible duplicates, and the table being huge
(several gigabytes) it takes dozens of minutes to execute a request.

Splits the query in batches, because Ecto seems to generate a
prepared statement such as `IN (?, ?, ?, ...) which eventually
reaches the limit of how many placeholders are allowed (65535).

fixes SURVEDA-MUMBAI-1E