Fix: refactor timeout/repeat in IVR simulator
Instead of handling the repetitions in the frontend and sending an
eventual STOP message to trick the simulator to breakoff, we send a
timeout message (similar to how the Verboice channel operates) to
tell the simulator (and its associated survey session) that we
didn't get a reply, which will trigger Surveda's retry mecanism,
which is to repeat the question twice before failing the session.

The behavior of the IVR simulator is now much closer to the reality.

It fixes the issue where a Contacted/Unresponsive simulation (no
answer) was flagged as a Responsive/Refused. It also avoids having a
STOP message to appear in the message history.