Disable pool timeout along with Repo's timeout
Generating the Interactions file is a long process that usually timeouts
for large surveys. But the process doesn't involve a single, long query
that times out - it instead performs a lot of queries, that time out the
pool of connections. So increasing the `Ask.Repo`'s `timeout` isn't
enough.

This PR makes the `DISABLE_REPO_TIMEOUT` flag to also disable the
`pool_timeout`.