Fix: handle completed verboice callbacks
Verboice has two callbacks: one to make progress in the call and another
for statuses. In the case where the progress callback ends the call, the
respondent won't have a session anymore when we receive the actual
status:completed callback.

This resulted in the channel broker getting stuck respondents when they
completed the survey (oops). This patch fixes the issue by immediately
notifying the channel broker with the completed state from the progress
callback.

Also a tiny performance improvement (only load the current mode, not the
whole session) and avoid relying on exceptions when trying to reach the
respondent's channel.