Speed up generation of respondent CSV files (#2253)
The bottleneck was the manipulation of datetimes to shift to the survey
scheduled timezone, as well as formatting those datetimes.

Also reduced the number of preloaded associations to avoid lots of
repeated requests, which may sound weird but we're talking batched
requests, which means that preloading associations will happen for
each batch. The actual resources are tiny (a few questionnaires, a few
channels at most) and we can preload them once, helping to reduce
lots of repeated, almost identical, SQL queries. It doesn't reduce the
performance of generating the CSV files but they significantly reduce
the amount of SQL queries .