Rework ChannelBroker GC to check the state of idle contacts (Verboice)
Instead of terminating active contacts that have been pending for a
while, we can ask the remote channel service for the actual state of the
IVR calls and SMS messages. We can then terminate them _if_ they're no
longer in queue or active.

This patch refactors the Channel Broker GC, and fixes the Verboice
channel to work with it (Nuntium is still pending).