Don't persist channel broker state to DB
Keeping the agent state in memory is enough. If a channel broker
crashes it shall be restarted with the last saved state in memory. If
Surveda crashing, everything will be restarted from zero anyway.

The worst case would be Surveda crashing in a loop, filling the external
channel up to capacity each time, which might lead to Surveda spamming
the external channels again. If we notice that happening in the wild
then we might reconsider.