Fix: internalize Ask.Runtime.Channel into ChannelBroker
We don't need to access it from outside of the ChannelBroker.

Now caches the Ask.Runtime.Channel instance into the channel broker's
state, and ask to refresh to oauth token when it's about to expire.
Since the token may only be refreshed once, we had to make sure that no
concurrent process would try to refresh it. Now that it's internalized
into the channel broker, it shouldn't be an issue.

Removes Ask.Runtime.ChannelAgent that failed regularly because of
Ecto.Adapters.SQL.Sandbox shared mode. esulted in a deadlock issue
identical to what happened when the channel broker tried to create a new
runtime channel (some process starts a transaction, while a nested call
tries to load the channel & token from the database -> deadlock).

Makes sure to load the initial runtime channel _before_ starting the
channel broker to avoid the above deadlock situation in the test suite.