Fix: check each channel status in a process
Restores the `spawn` call to check each channel status in its own
process, allowing parallel checks and avoiding a crash while checking a
channel to block checking the status of the following channels (AFAIK).

The change was only introduced to fix a test, but I can't replicate the
issue anymore. See:
https://github.com/instedd/surveda/pull/2127/commits/a58fbde4ba6630b2794f217a0f8443c9bc4c3baf