Increase DB chunk size for respondent files
Respondent files are usually large (Interactions files can grow up to 1M
rows), and the "low" limit in queries made the DB work much more than
needed (we've observed 99% CPU usage in the mysqld process when
generating a 1M-rows interactions file with 1000 rows per query).

Increasing this limit makes the app generate less queries to the DB,
effectively driving the CPU usage down to about 30% instead.

There's probably more room for improvement (the generation of the file
is still CPU-bound instead of network-bound), but that's on the app
itself - we should profile the app's code to further improve the
performance.

See #2350
See #2359

Co-authored-by: Gustavo Giráldez <ggiraldez@manas.tech>