Fix: filter was not updated in file status queries
Due to the way the `q` variable was bound in the previous
implementation of the `setInterval` call, we were always querying the
file status using the filter (or lack thereof) that was in place in the
first show of the modal.

We now get the current value of `q` at the time in which the refresh
function is called.

See #2350