Make respondent session optional in tests
With the changes in b5b7ee409637859419f1441c12353681e61e4a89, the
ChannelBroker now makes it mandatory for a respondent to have a session
already initialized. That broke a huge amount of tests.

We know we should decouple this in the code. Until we do that, this
commit supports session-less respondents only when in test mode.

See #2368
See #2382