Fix: tests that integrate with other processes
We were observing tests failing (or logging crashes) because some
process was still trying to process messages from a few tests before.

According to https://hexdocs.pm/elixir/docs-tests-and-with.html it's
recommended to stop the application, then start it again for each
integration test in order to clear every process' state and mailbox.

Co-authored-by: Ana Pérez Ghiglia <aghiglia@manas.tech>